Kanas: power: Port Android N's powerhal based on the scx35-common version [WIP]
I'm still plagued by the problem with the scx35-common version
specially when battery save mode is triggered once. There are
a couple of problems here and there like:
  An file descriptor leak
  Two counts of memory leak
  A broken inlined atoi implementation

In the case of battery saver, since this is specifically made for the
interactive governor, the code expects interactive governor knobs.
This device doesn't use the interactive governor.
In the Stock Rom, IIRC, only have like 2 governors built in.

The trigger for battery saver correctly limits the maximum CPU freq
but incorrectly lifts it since it depends on the hispeed_freq knob
that does not exists in the sprdemand governor. The result,
you're out of the battery saver mode but the CPU clocks are still limited.

This version is ported from the Android N version unto the scx35-common
version with some few changes.
  It's now possible to change governors on runtime and have the PowerHAL
  acknowledge that.
  LCD refresh rate, if allowed by the kernel, can be brought down to 33 fps.
That's all.

As a refresher, These are what was added in Android N version:
  + LCD Refresh rate is limited to 45 fps when going below balance profile
  + 2 more profiles
  + CPULimit support
  + support to my gpu core scaler algorithm
  + mimics the option to limit CPU frequency found in the stock rom

Change-Id: I50aa4895ddb850d060b38e239f1f5426a38dd54b
Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>